+++ Content/Items/Weapons/Ranged/SuperSniper/SuperSniperHoldout.cs
@@ -23,6 +23,7 @@ namespace SuperfluityTwo.Content.Items.Weapons.Ranged.SuperSniper
 
         ref float useTimer => ref Projectile.ai[0];
         ref float dustTimer => ref Projectile.ai[1];
+        ref float recoilTimer => ref Projectile.ai[2];
 
         public override void Load()
         {
@@ -93,33 +94,75 @@ namespace SuperfluityTwo.Content.Items.Weapons.Ranged.SuperSniper
         private void HandleFiring()
         {
             Player player = Main.player[Projectile.owner];
-            float chargeRatio = (GetDefaultUseTimer() - useTimer) / GetDefaultUseTimer();
+            float chargeRatio = 1f - (useTimer / GetDefaultUseTimer());
             Vector2 dir = Projectile.velocity.SafeNormalize(-Vector2.UnitY);
-            dustTimer += 30.0f * chargeRatio;
-            if (useTimer <= 0 && player.whoAmI == Main.myPlayer && player.PickAmmo(player.HeldItem, out int proj, out float speed, out int damage, out float kB, out int itemID))
+            dustTimer += chargeRatio;
+            if (useTimer <= 0)
             {
-                Projectile.NewProjectile(
-                    player.GetSource_FromThis(),
-                    Main.MouseWorld,
-                    dir * speed,
-                    ModContent.ProjectileType<SuperSniperReticle>(),
-                    damage,
-                    kB,
-                    player.whoAmI
+                if (!player.PickAmmo(player.HeldItem, out int proj, out float speed, out int damage, out float kB, out int itemID))
+                {
+                    Projectile.Kill();
+                    return;
+                }
+
+                if (player.whoAmI == Main.myPlayer)
+                {
+                    Projectile.NewProjectileDirect(
+                        player.GetSource_FromThis(),
+                        Main.MouseWorld,
+                        dir * speed,
+                        ModContent.ProjectileType<SuperSniperReticle>(),
+                        damage,
+                        kB,
+                        player.whoAmI
+                    ).CritChance = player.HeldItem.crit;
+                }
+
+                Vector2 muzzleLocation = new Vector2(40, -2 * Projectile.direction).RotatedBy(Projectile.rotation);
+
+                for (int i = 0; i < 5; i++)
+                {
+                    Dust.NewDustPerfect(
+                        Position: Projectile.Center + muzzleLocation,
+                        Type: DustID.Smoke,
+                        Velocity: (1.75f + Main.rand.NextFloat()) * dir.RotatedByRandom(MathHelper.PiOver4 * 0.5f),
+                        Scale: 2f
+                    ).noGravity = true;
+                    Dust.NewDustPerfect(
+                        Position: Projectile.Center + muzzleLocation,
+                        Type: DustID.Torch,
+                        Velocity: (1.75f + Main.rand.NextFloat()) * dir.RotatedByRandom(MathHelper.PiOver4 * 0.5f),
+                        Scale: 2f
+                    ).noGravity = true;
+                }
+                Gore gore = Gore.NewGorePerfect(
+                    source: Projectile.GetSource_FromThis(),
+                    Position: Projectile.Center + muzzleLocation,
+                    Velocity: 0.9f * dir.RotatedByRandom(MathHelper.PiOver4 * 0.5f),
+                    Type: Main.rand.Next(GoreID.Smoke1, GoreID.Smoke3 + 1),
+                    Scale: 0.8f
                 );
+                gore.position -= new Vector2(gore.Width / 2, gore.Height / 2);
+                gore.position += dir * (gore.Width / 2);
+                gore.rotation = Main.rand.NextFloat(MathHelper.TwoPi);
+                gore.timeLeft /= 2;
+
                 ResetUseTimer();
+                recoilTimer = 10;
                 Projectile.netUpdate = true;
             }
             while (dustTimer > 0)
             {
                 Dust.NewDustPerfect(
-                    Projectile.Center + dir * 30,
+                    Projectile.Center + new Vector2(40, -2 * Projectile.direction).RotatedBy(Projectile.rotation),
                     DustID.GoldFlame,
-                    dir + 0.3f * new Vector2(Main.rand.Next() - 0.5f, Main.rand.Next() - 0.5f)
+                    (1.5f + chargeRatio) * dir.RotatedByRandom(MathHelper.PiOver4 * 0.5f),
+                    Scale: chargeRatio + 0.5f
                 ).noGravity = true;
                 dustTimer--;
             }
             useTimer--;
+            if (recoilTimer > 0) recoilTimer--;
         }
 
         private void ResetUseTimer()
@@ -156,8 +199,8 @@ namespace SuperfluityTwo.Content.Items.Weapons.Ranged.SuperSniper
             float rotation = Projectile.velocity.ToRotation();
             Projectile.spriteDirection = Projectile.direction;
 
-            Vector2 holdoutOffset = new Vector2(0, -34);
-            holdoutOffset.Y += -0.5f;
+            Vector2 holdoutOffset = new Vector2(3, -35);
+            holdoutOffset.Y += recoilTimer * 0.2f;
 
             Projectile.Center = playerHandPos + new Vector2(holdoutOffset.X * Projectile.direction, holdoutOffset.Y).RotatedBy(Projectile.rotation);
             Projectile.rotation = rotation;
@@ -176,7 +219,11 @@ namespace SuperfluityTwo.Content.Items.Weapons.Ranged.SuperSniper
             Texture2D texture = texAsset.Value;
             int frameHeight = texture.Height;
             int spriteSheetOffset = frameHeight * Projectile.frame;
+            
             Vector2 sheetInsertPosition = (Projectile.Center + Vector2.UnitY * Projectile.gfxOffY - Main.screenPosition).Floor();
+            
+            Vector2 correction = new Vector2(0, -Projectile.direction + 1).RotatedBy(Projectile.rotation);
+            sheetInsertPosition += correction;
 
             Color drawColor = Projectile.GetAlpha(lightColor);
             Main.EntitySpriteDraw(texture, sheetInsertPosition, new Rectangle?(new Rectangle(0, spriteSheetOffset, texture.Width, frameHeight - 2)), drawColor, Projectile.rotation, new Vector2(texture.Width / 2f, frameHeight / 2f), Projectile.scale, effects, 0f);